update: Routerをコンポーネント化
