+++ front/src/App.tsx
@@ -1,101 +1,19 @@
-import React, { useState, useEffect, createContext } from "react";
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-  Redirect,
-} from "react-router-dom";
+import React from "react";
+import { Router } from "router/Router";
+import { BrowserRouter } from "react-router-dom";
 
-import { CommonLayout } from "components/templates/CommonLayout";
-import { Home } from "components/pages/Home";
-import { SignUp } from "components/pages/SignUp";
-import { SignIn } from "components/pages/SignIn";
+import { CssBaseline, ThemeProvider } from "@mui/material";
 
-import { getCurrentUser } from "lib/api/auth/getCurrentUser";
-import { User } from "types/api/user";
-
-// グローバルで扱う変数・関数
-export const AuthContext = createContext(
-  {} as {
-    loading: boolean;
-    setLoading: React.Dispatch<React.SetStateAction<boolean>>;
-    isSignedIn: boolean;
-    setIsSignedIn: React.Dispatch<React.SetStateAction<boolean>>;
-    currentUser: User | undefined;
-    setCurrentUser: React.Dispatch<React.SetStateAction<User | undefined>>;
-  }
-);
+import theme from "theme/theme";
 
 function App() {
-  const [loading, setLoading] = useState<boolean>(true);
-  const [isSignedIn, setIsSignedIn] = useState<boolean>(false);
-  const [currentUser, setCurrentUser] = useState<User | undefined>();
-
-  // 認証済みのユーザーがいるかどうかチェック
-  // 確認できた場合はそのユーザーの情報を取得
-  const handleGetCurrentUser = async () => {
-    try {
-      const res = await getCurrentUser();
-      console.log(res);
-
-      if (res?.status === 200) {
-        setIsSignedIn(true);
-        setCurrentUser(res?.data.currentUser);
-      } else {
-        alert("ユーザーがいません");
-      }
-    } catch (err) {
-      console.log(err);
-    }
-
-    setLoading(false);
-  };
-
-  useEffect(() => {
-    handleGetCurrentUser();
-  }, [setCurrentUser]);
-
-  // ユーザーが認証済みかどうかでルーティングを決定
-  // 未認証だった場合は「/signin」ページに促す
-  const Private = ({ children }: { children: React.ReactElement }) => {
-    if (!loading) {
-      if (isSignedIn) {
-        return children;
-      } else {
-        return <Redirect to="/signin" />;
-      }
-    } else {
-      return <></>;
-    }
-  };
-
   return (
-    <Router>
-      {/* routerフォルダとしに切り分けたい */}
-      {/* AuthProviderとしてproviderフォルダに切り分けたい */}
-      <AuthContext.Provider
-        value={{
-          loading,
-          setLoading,
-          isSignedIn,
-          setIsSignedIn,
-          currentUser,
-          setCurrentUser,
-        }}
-      >
-        <CommonLayout>
-          <Switch>
-            <Route exact path="/signup" component={SignUp} />
-            <Route exact path="/signin" component={SignIn} />
-            <Private>
-              <Switch>
-                <Route exact path="/" component={Home} />
-              </Switch>
-            </Private>
-          </Switch>
-        </CommonLayout>
-      </AuthContext.Provider>
-    </Router>
+    <ThemeProvider theme={theme}>
+      <CssBaseline />
+      <BrowserRouter>
+        <Router />
+      </BrowserRouter>
+    </ThemeProvider>
   );
 }
